requirements.txt
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
dnspython==2.4.2
executing==1.2.0
Flask==2.3.2
Flask-Session==0.5.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.4.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
